+++ src-plugin/Plugin/Models/ArenaPlayerModel.cs
@@ -242,7 +242,7 @@ public class ArenaPlayer
 		ChatMenu weaponPreferenceMenu = new ChatMenu(Localizer.ForPlayer(Controller, "k4.menu.weaponpref.title"));
 		foreach (WeaponType weaponType in Enum.GetValues(typeof(WeaponType)))
 		{
-			if (weaponType == WeaponType.Unknown)
+			if (weaponType == WeaponType.Unknown || !IsAllowedWeaponType(weaponType))
 				continue;
 			weaponPreferenceMenu.AddMenuOption(Localizer.ForPlayer(Controller, $"k4.rounds.{weaponType.ToString().ToLower()}"),
 				(player, option) =>
@@ -257,11 +257,14 @@ public class ArenaPlayer
 	private void ShowCenterWeaponPreferenceMenu()
 	{
 		var items = new List<MenuItem>();
+		var values = new Dictionary<int, WeaponType>();
+		int count = 0;
 		foreach (WeaponType weaponType in Enum.GetValues(typeof(WeaponType)))
 		{
-			if (weaponType == WeaponType.Unknown)
+			if (weaponType == WeaponType.Unknown || !IsAllowedWeaponType(weaponType))
 				continue;
 			items.Add(new MenuItem(MenuItemType.Button, [new MenuValue($"{Localizer.ForPlayer(Controller, $"k4.rounds.{weaponType.ToString().ToLower()}")}")]));
+			values.Add(count++, weaponType);
 		}
 
 		Plugin.Menu?.ShowScrollableMenu(Controller, Localizer.ForPlayer(Controller, "k4.menu.weaponpref.title"), items, (buttons, menu, selected) =>
@@ -277,10 +280,24 @@ public class ArenaPlayer
 			if (selected == null) return;
 			if (buttons == MenuButtons.Select)
 			{
-				WeaponType selectedWeaponType = (WeaponType)menu.Option;
+				WeaponType selectedWeaponType = values[menu.Option];
 				ShowWeaponSubPreferenceMenu(selectedWeaponType);
 			}
 		}, false, Config.CommandSettings.FreezeInMenu, disableDeveloper: Config.CommandSettings.ShowMenuCredits);
+	}
+
+	private bool IsAllowedWeaponType(WeaponType weaponType)
+	{
+		return weaponType switch
+		{
+			WeaponType.Rifle => Config.AllowedWeaponPreferences.Rifle,
+			WeaponType.Sniper => Config.AllowedWeaponPreferences.Sniper,
+			WeaponType.SMG => Config.AllowedWeaponPreferences.SMG,
+			WeaponType.LMG => Config.AllowedWeaponPreferences.LMG,
+			WeaponType.Shotgun => Config.AllowedWeaponPreferences.Shotgun,
+			WeaponType.Pistol => Config.AllowedWeaponPreferences.Pistol,
+			_ => false
+		};
 	}
 
 	public void ShowWeaponSubPreferenceMenu(WeaponType weaponType)